backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==22.0.0
psycopg2-binary==2.9.9

# Performance optimization dependencies (optional)
redis==5.0.1
sqlalchemy-utils==0.41.1
